+++ python/instrumentation/openinference-instrumentation-llama-index/src/openinference/instrumentation/llama_index/_handler.py
@@ -16,6 +16,7 @@ from typing import (
     TYPE_CHECKING,
     Any,
     AsyncGenerator,
+    Callable,
     DefaultDict,
     Dict,
     Generator,
@@ -48,6 +49,9 @@ from llama_index.core.base.llms.types import (
     ChatMessage,
     ChatResponse,
     CompletionResponse,
+    ContentBlock,
+    ImageBlock,
+    TextBlock,
 )
 from llama_index.core.base.response.schema import (
     RESPONSE_TYPE,
@@ -125,6 +129,7 @@ from openinference.semconv.trace import (
     OpenInferenceSpanKindValues,
     RerankerAttributes,
     SpanAttributes,
+    ToolAttributes,
     ToolCallAttributes,
 )
 
@@ -223,6 +228,13 @@ class _Span(BaseSpan):
         return set_span_in_context(self._otel_span)
 
     def process_input(self, instance: Any, bound_args: inspect.BoundArguments) -> None:
+        from llama_index.core.llms.function_calling import FunctionCallingLLM
+
+        if isinstance(instance, FunctionCallingLLM) and isinstance(
+            (tools := bound_args.kwargs.get("tools")), Iterable
+        ):
+            for i, tool in enumerate(tools):
+                self[f"{LLM_TOOLS}.{i}.{TOOL_JSON_SCHEMA}"] = safe_json_dumps(tool)
         try:
             self[INPUT_VALUE] = safe_json_dumps(bound_args.arguments, cls=_Encoder)
             self[INPUT_MIME_TYPE] = JSON
@@ -252,9 +264,11 @@ class _Span(BaseSpan):
             try:
                 self[OUTPUT_VALUE] = result.model_dump_json(exclude_unset=True)
                 self[OUTPUT_MIME_TYPE] = JSON
-            except BaseException as e:
-                logger.exception(str(e))
-                pass
+            except Exception:
+                try:
+                    self[OUTPUT_VALUE] = repr(result)
+                except Exception:
+                    pass
         else:
             try:
                 self[OUTPUT_VALUE] = safe_json_dumps(result, cls=_Encoder)
@@ -528,23 +542,20 @@ class _Span(BaseSpan):
         self._process_response_text_type(event.response)
 
     def _extract_token_counts(self, response: Union[ChatResponse, CompletionResponse]) -> None:
-        if (
-            (raw := getattr(response, "raw", None))
-            and hasattr(raw, "get")
-            and (usage := raw.get("usage"))
-        ):
-            for k, v in _get_token_counts(usage):
-                self[k] = v
-        if (
-            (raw := getattr(response, "raw", None))
-            and (model_extra := getattr(raw, "model_extra", None))
-            and hasattr(model_extra, "get")
-            and (x_groq := model_extra.get("x_groq"))
-            and hasattr(x_groq, "get")
-            and (usage := x_groq.get("usage"))
-        ):
-            for k, v in _get_token_counts(usage):
-                self[k] = v
+        if raw := getattr(response, "raw", None):
+            usage = raw.get("usage") if isinstance(raw, Mapping) else getattr(raw, "usage", None)
+            if usage:
+                for k, v in _get_token_counts(usage):
+                    self[k] = v
+            if (
+                (model_extra := getattr(raw, "model_extra", None))
+                and hasattr(model_extra, "get")
+                and (x_groq := model_extra.get("x_groq"))
+                and hasattr(x_groq, "get")
+                and (usage := x_groq.get("usage"))
+            ):
+                for k, v in _get_token_counts(usage):
+                    self[k] = v
         # Look for token counts in additional_kwargs of the completion payload
         # This is needed for non-OpenAI models
         if additional_kwargs := getattr(response, "additional_kwargs", None):
@@ -568,15 +579,16 @@ class _Span(BaseSpan):
     ) -> None:
         for i, message in enumerate(messages):
             self[f"{prefix}.{i}.{MESSAGE_ROLE}"] = message.role.value
-            if content := message.content:
-                if isinstance(content, str):
-                    self[f"{prefix}.{i}.{MESSAGE_CONTENT}"] = str(content)
-                elif is_iterable_of(content, dict):
-                    for j, c in list(enumerate(content)):
-                        for key, value in self._get_attributes_from_message_content(
-                            c,
-                        ):
-                            self[f"{prefix}.{i}.{MESSAGE_CONTENTS}.{j}.{key}"] = value
+            blocks = message.blocks
+            if len(blocks) == 1 and isinstance(blocks[0], TextBlock):
+                self[f"{prefix}.{i}.{MESSAGE_CONTENT}"] = blocks[0].text
+            else:
+                for j, block in enumerate(blocks):
+                    for k, v in _get_attributes_from_content_block(
+                        block,
+                        prefix=f"{prefix}.{i}.{MESSAGE_CONTENTS}.{j}.",
+                    ):
+                        self[k] = v
             additional_kwargs = message.additional_kwargs
             if name := additional_kwargs.get("name"):
                 self[f"{prefix}.{i}.{MESSAGE_NAME}"] = name
@@ -652,6 +664,40 @@ class _Span(BaseSpan):
             yield f"{ImageAttributes.IMAGE_URL}", url
 
 
+def _get_attributes_from_content_block(
+    obj: ContentBlock,
+    prefix: str,
+) -> Iterator[Tuple[str, AttributeValue]]:
+    if isinstance(obj, TextBlock):
+        yield from _get_attributes_from_text_block(obj, prefix=prefix)
+    elif isinstance(obj, ImageBlock):
+        yield from _get_attributes_from_image_block(obj, prefix=prefix)
+
+
+def _get_attributes_from_text_block(
+    obj: TextBlock,
+    prefix: str = "",
+) -> Iterator[Tuple[str, AttributeValue]]:
+    yield f"{prefix}{MESSAGE_CONTENT_TYPE}", "text"
+    yield f"{prefix}{MESSAGE_CONTENT_TEXT}", obj.text
+
+
+def _get_attributes_from_image_block(
+    obj: ImageBlock,
+    prefix: str = "",
+) -> Iterator[Tuple[str, AttributeValue]]:
+    if obj.image and obj.image_mimetype:
+        url = f"data:{obj.image_mimetype};base64,{obj.image.decode()}"
+        yield f"{prefix}{MESSAGE_CONTENT_IMAGE}.{IMAGE_URL}", url
+        yield f"{prefix}{MESSAGE_CONTENT_TYPE}", "image"
+    elif obj.url:
+        yield f"{prefix}{MESSAGE_CONTENT_IMAGE}.{IMAGE_URL}", str(obj.url)
+        yield f"{prefix}{MESSAGE_CONTENT_TYPE}", "image"
+    elif obj.path:
+        yield f"{prefix}{MESSAGE_CONTENT_IMAGE}.{IMAGE_URL}", str(obj.path)
+        yield f"{prefix}{MESSAGE_CONTENT_TYPE}", "image"
+
+
 END_OF_QUEUE = None
 
 
@@ -713,11 +759,24 @@ class _ExportQueue:
 
 class _SpanHandler(BaseSpanHandler[_Span], extra="allow"):
     _otel_tracer: Tracer = PrivateAttr()
+    _separate_trace_from_runtime_context: bool = PrivateAttr()
     _export_queue: _ExportQueue = PrivateAttr()
 
-    def __init__(self, tracer: Tracer) -> None:
+    def __init__(
+        self,
+        tracer: Tracer,
+        separate_trace_from_runtime_context: bool = False,
+    ) -> None:
+        """Initialize the span handler.
+
+        Args:
+            tracer (trace_api.Tracer): The OpenTelemetry tracer for creating spans.
+            separate_trace_from_runtime_context (bool): When True, always start a new trace for each
+                span without a parent, isolating it from any existing trace in the runtime context.
+        """
         super().__init__()
         self._otel_tracer = tracer
+        self._separate_trace_from_runtime_context = separate_trace_from_runtime_context
         self._export_queue = _ExportQueue()
 
     def new_span(
@@ -737,7 +796,11 @@ class _SpanHandler(BaseSpanHandler[_Span], extra="allow"):
             name=id_.partition("-")[0],
             start_time=time_ns(),
             attributes=dict(get_attributes_from_context()),
-            context=(parent.context if parent else None),
+            context=(
+                parent.context
+                if parent
+                else (context_api.Context() if self._separate_trace_from_runtime_context else None)
+            ),
         )
         span = _Span(
             otel_span=otel_span,
@@ -803,9 +866,9 @@ class _SpanHandler(BaseSpanHandler[_Span], extra="allow"):
 class EventHandler(BaseEventHandler, extra="allow"):
     _span_handler: _SpanHandler = PrivateAttr()
 
-    def __init__(self, tracer: Tracer) -> None:
+    def __init__(self, span_handler: _SpanHandler) -> None:
         super().__init__()
-        self._span_handler = _SpanHandler(tracer=tracer)
+        self._span_handler = span_handler
 
     def handle(self, event: BaseEvent, **kwargs: Any) -> Any:
         if context_api.get_value(_SUPPRESS_INSTRUMENTATION_KEY):
@@ -854,39 +917,93 @@ def _get_token_counts(usage: Union[object, Mapping[str, Any]]) -> Iterator[Tuple
 
 
 def _get_token_counts_from_object(usage: object) -> Iterator[Tuple[str, Any]]:
-    if (prompt_tokens := getattr(usage, "prompt_tokens", None)) is not None:
-        try:
-            yield LLM_TOKEN_COUNT_PROMPT, int(prompt_tokens)
-        except BaseException:
-            pass
-    if (completion_tokens := getattr(usage, "completion_tokens", None)) is not None:
-        try:
-            yield LLM_TOKEN_COUNT_COMPLETION, int(completion_tokens)
-        except BaseException:
-            pass
-    if (total_tokens := getattr(usage, "total_tokens", None)) is not None:
-        try:
-            yield LLM_TOKEN_COUNT_TOTAL, int(total_tokens)
-        except BaseException:
-            pass
+    def get_value(obj: object, key: str) -> Any:
+        return getattr(obj, key, None)
+
+    yield from _get_token_counts_impl(usage, get_value)
 
 
 def _get_token_counts_from_mapping(
     usage_mapping: Mapping[str, Any],
 ) -> Iterator[Tuple[str, Any]]:
-    if (prompt_tokens := usage_mapping.get("prompt_tokens")) is not None:
+    def get_value(obj: Mapping[str, Any], key: str) -> Any:
+        return obj.get(key)
+
+    yield from _get_token_counts_impl(usage_mapping, get_value)
+
+
+def _get_token_counts_impl(
+    usage: Union[object, Mapping[str, Any]], get_value: Callable[[Any, str], Any]
+) -> Iterator[Tuple[str, Any]]:
+    # OpenAI
+    if (prompt_tokens := get_value(usage, "prompt_tokens")) is not None:
         try:
             yield LLM_TOKEN_COUNT_PROMPT, int(prompt_tokens)
         except BaseException:
             pass
-    if (completion_tokens := usage_mapping.get("completion_tokens")) is not None:
+    if (prompt_token_details := get_value(usage, "prompt_tokens_details")) is not None:
+        if (cached_tokens := get_value(prompt_token_details, "cached_tokens")) is not None:
+            try:
+                yield LLM_TOKEN_COUNT_PROMPT_DETAILS_CACHE_READ, int(cached_tokens)
+            except BaseException:
+                pass
+        if (audio_tokens := get_value(prompt_token_details, "audio_tokens")) is not None:
+            try:
+                yield LLM_TOKEN_COUNT_PROMPT_DETAILS_AUDIO, int(audio_tokens)
+            except BaseException:
+                pass
+    if (completion_tokens := get_value(usage, "completion_tokens")) is not None:
         try:
             yield LLM_TOKEN_COUNT_COMPLETION, int(completion_tokens)
         except BaseException:
             pass
-    if (total_tokens := usage_mapping.get("total_tokens")) is not None:
+    if (completion_tokens_details := get_value(usage, "completion_tokens_details")) is not None:
+        if (
+            reasoning_tokens := get_value(completion_tokens_details, "reasoning_tokens")
+        ) is not None:
+            try:
+                yield LLM_TOKEN_COUNT_COMPLETION_DETAILS_REASONING, int(reasoning_tokens)
+            except BaseException:
+                pass
+        if (
+            completion_audio_tokens := get_value(
+                completion_tokens_details, "audio_tokens" if isinstance(usage, object) else "audio"
+            )
+        ) is not None:
+            try:
+                yield LLM_TOKEN_COUNT_COMPLETION_DETAILS_AUDIO, int(completion_audio_tokens)
+            except BaseException:
+                pass
+    if (total_tokens := get_value(usage, "total_tokens")) is not None:
         try:
             yield LLM_TOKEN_COUNT_TOTAL, int(total_tokens)
+        except BaseException:
+            pass
+
+    # Anthropic
+    if (output_tokens := get_value(usage, "output_tokens")) is not None:
+        try:
+            yield LLM_TOKEN_COUNT_COMPLETION, int(output_tokens)
+        except BaseException:
+            pass
+    if (cache_creation_input_tokens := get_value(usage, "cache_creation_input_tokens")) is not None:
+        try:
+            yield LLM_TOKEN_COUNT_PROMPT_DETAILS_CACHE_WRITE, int(cache_creation_input_tokens)
+        except BaseException:
+            pass
+    if (cache_read_input_tokens := get_value(usage, "cache_read_input_tokens")) is not None:
+        try:
+            yield LLM_TOKEN_COUNT_PROMPT_DETAILS_CACHE_READ, int(cache_read_input_tokens)
+        except BaseException:
+            pass
+    if (input_tokens := get_value(usage, "input_tokens")) is not None:
+        try:
+            input_tokens = int(input_tokens)
+            if cache_creation_input_tokens is not None:
+                input_tokens += int(cache_creation_input_tokens)
+            if cache_read_input_tokens is not None:
+                input_tokens += int(cache_read_input_tokens)
+            yield LLM_TOKEN_COUNT_PROMPT, input_tokens
         except BaseException:
             pass
 
@@ -1042,8 +1159,19 @@ LLM_PROMPTS = SpanAttributes.LLM_PROMPTS
 LLM_PROMPT_TEMPLATE = SpanAttributes.LLM_PROMPT_TEMPLATE
 LLM_PROMPT_TEMPLATE_VARIABLES = SpanAttributes.LLM_PROMPT_TEMPLATE_VARIABLES
 LLM_TOKEN_COUNT_COMPLETION = SpanAttributes.LLM_TOKEN_COUNT_COMPLETION
+LLM_TOKEN_COUNT_COMPLETION_DETAILS_AUDIO = SpanAttributes.LLM_TOKEN_COUNT_COMPLETION_DETAILS_AUDIO
+LLM_TOKEN_COUNT_COMPLETION_DETAILS_REASONING = (
+    SpanAttributes.LLM_TOKEN_COUNT_COMPLETION_DETAILS_REASONING
+)
 LLM_TOKEN_COUNT_PROMPT = SpanAttributes.LLM_TOKEN_COUNT_PROMPT
+LLM_TOKEN_COUNT_PROMPT_DETAILS_AUDIO = SpanAttributes.LLM_TOKEN_COUNT_PROMPT_DETAILS_AUDIO
+LLM_TOKEN_COUNT_PROMPT_DETAILS_CACHE_READ = SpanAttributes.LLM_TOKEN_COUNT_PROMPT_DETAILS_CACHE_READ
+LLM_TOKEN_COUNT_PROMPT_DETAILS_CACHE_WRITE = (
+    SpanAttributes.LLM_TOKEN_COUNT_PROMPT_DETAILS_CACHE_WRITE
+)
+
 LLM_TOKEN_COUNT_TOTAL = SpanAttributes.LLM_TOKEN_COUNT_TOTAL
+LLM_TOOLS = SpanAttributes.LLM_TOOLS
 MESSAGE_CONTENT = MessageAttributes.MESSAGE_CONTENT
 MESSAGE_CONTENTS = MessageAttributes.MESSAGE_CONTENTS
 MESSAGE_CONTENT_TYPE = MessageContentAttributes.MESSAGE_CONTENT_TYPE
@@ -1071,6 +1199,8 @@ TOOL_CALL_FUNCTION_NAME = ToolCallAttributes.TOOL_CALL_FUNCTION_NAME
 TOOL_DESCRIPTION = SpanAttributes.TOOL_DESCRIPTION
 TOOL_NAME = SpanAttributes.TOOL_NAME
 TOOL_PARAMETERS = SpanAttributes.TOOL_PARAMETERS
+
+TOOL_JSON_SCHEMA = ToolAttributes.TOOL_JSON_SCHEMA
 
 JSON = OpenInferenceMimeTypeValues.JSON.value