+++ python/instrumentation/openinference-instrumentation-openai-agents/src/openinference/instrumentation/openai_agents/package.py
@@ -1,2 +1,2 @@
-_instruments = ("openai-agents >= 0.0.3",)
+_instruments = ("openai-agents >= 0.0.7",)
 _supports_metrics = False