+++ js/packages/openinference-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @arizeai/openinference-core
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [ae5cd15]
+  - @arizeai/openinference-semantic-conventions@1.1.0
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [c4e2252]
+  - @arizeai/openinference-semantic-conventions@1.0.1
+
 ## 1.0.0
 
 ### Major Changes